+++ web_ui/frontend/src/App.tsx
@@ -9,7 +9,6 @@ import { getApiErrorMessage, loadTub } from './services/api';
 
 const TrainerPage = React.lazy(() => import('./pages/TrainerPage').then((module) => ({ default: module.TrainerPage })));
 const DrivePage = React.lazy(() => import('./pages/DrivePage').then((module) => ({ default: module.DrivePage })));
-const CalibratePage = React.lazy(() => import('./pages/CalibratePage').then((module) => ({ default: module.CalibratePage })));
 const PilotArenaPage = React.lazy(() => import('./pages/PilotArenaPage').then((module) => ({ default: module.PilotArenaPage })));
 const CarConnectorPage = React.lazy(() => import('./pages/CarConnectorPage').then((module) => ({ default: module.CarConnectorPage })));
 
@@ -103,7 +102,6 @@ function AppShell() {
             <Route path="/" element={<TubManagerPage />} />
             <Route path="/trainer" element={<TrainerPage />} />
             <Route path="/drive" element={<DrivePage />} />
-            <Route path="/calibrate" element={<CalibratePage />} />
             <Route path="/pilot" element={<PilotArenaPage />} />
             <Route path="/connector" element={<CarConnectorPage />} />
           </Routes>